--- src/App.vue
@@ -1,82 +1,0 @@
-<template>
-  <div id="app">
-    <router-view />
-  </div>
-</template>
-
-<script>
-import { FindInPage } from "electron-find";
-import { ipcRenderer } from "electron";
-
-const remote = require("@electron/remote");
-let webContent = remote.getCurrentWebContents();
-let findInPage = new FindInPage(webContent, {
-  boxBgColor: "#333",
-  boxShadowColor: "#000",
-  inputColor: "#aaa",
-  inputBgColor: "#222",
-  inputFocusColor: "#555",
-  textColor: "#aaa",
-  textHoverBgColor: "#555",
-  caseSelectedColor: "#555",
-});
-export default {
-  data() {
-    return {
-      catch_components: [],
-      editableTabsValue: "2",
-      editableTabs: [],
-      tabIndex: 2,
-      ccc: "",
-    };
-  },
-  mounted() {
-    ipcRenderer.on("asynchronous-reply", function () {
-      // 接收到Main进程返回的消息
-      findInPage.openFindWindow();
-    });
-
-  },
-  computed: {},
-  methods: {},
-};
-</script>
-<style>
-#app {
-  font-family: Avenir, Helvetica, Arial, sans-serif;
-  -webkit-font-smoothing: antialiased;
-  -moz-osx-font-smoothing: grayscale;
-
-  color: #2c3e50;
-}
-
-#nav {
-  text-align: center;
-}
-
-body {
-  position: absolute;
-  top: 0px;
-  bottom: 0px;
-  left: 0px;
-  right: 0px;
-}
-
-
-.notSelect {
-  /* 不让用户选中 */
-  -webkit-touch-callout: none; /* iOS Safari */
-  -webkit-user-select: none; /* Chrome/Safari/Opera */
-  -khtml-user-select: none; /* Konqueror */
-  -moz-user-select: none; /* Firefox */
-  -ms-user-select: none; /* Internet Explorer/Edge */
-  user-select: none; /* Non-prefixed version, currently
-  not supported by any browser */
-}
-
-.el-tabs__item:focus.is-active.is-focus:not(:active) {
-	-webkit-box-shadow: none !important;
-	box-shadow: none !important;
-}
-
-</style>
