+++ src/renderer/assets/style/global.css
@@ -1,0 +1,14 @@
+
+
+/* 
+    element-ui dialog的内容内边距
+*/
+.el-dialog .el-dialog__body {
+    padding-top: 0px;
+    padding-left: 10px;
+    padding-right: 10px;
+    padding-bottom: 10px;
+  }
+
+
+
