+++ src/renderer/views/welcome.vue
@@ -2,13 +2,20 @@
   <div class="welcomeContainer notSelect">
     <div class="contentContainer">
       <div class="imgContainer">
-        <img src="../../assets/icon.png" width="120" height="120" />
+        <img src="../assets/icon.png" width="120" height="120" />
       </div>
       <div class="descContainer">
         <span class="title">{{$t("welcome.productName")}}</span>
       </div>
       <div>
         <span class="version">{{$t("welcome.version")}}{{version}} </span>
+      </div>
+      <div>
+
+        <iframe src="https://ghbtns.com/github-btn.html?user=qiaolin-li&repo=dubbo-desktop-manager&type=watch&count=true" frameborder="0" scrolling="0" width="120" height="20" title="GitHub"></iframe>
+
+        <iframe src="https://ghbtns.com/github-btn.html?user=qiaolin-li&repo=dubbo-desktop-manager&type=fork&count=true" frameborder="0" scrolling="0" width="120" height="20" title="GitHub"></iframe>
+        <iframe src="https://ghbtns.com/github-btn.html?user=qiaolin-li&repo=dubbo-desktop-manager&type=watch&count=true&v=2" frameborder="0" scrolling="0" width="120" height="20" title="GitHub"></iframe>
       </div>
     </div>
 
@@ -17,6 +24,8 @@
 
 <script>
 import pkg from "../../../package.json";
+const { shell } = require('electron');
+
 export default {
   data() {
     return {
@@ -28,6 +37,21 @@ export default {
   created() {
   },
   mounted() {
+    let iframeList = document.getElementsByTagName("iframe");
+
+    for (let i = 0; i < iframeList.length; i++) {
+      const iframe = iframeList[i];
+      iframe.onload = function () {
+        const links = iframe.contentWindow.document.querySelectorAll('a[href]');
+        links.forEach(link => {
+          link.addEventListener('click', e => {
+            const url = link.getAttribute('href');
+            e.preventDefault();
+            shell.openExternal(url);
+          });
+        });
+      }
+    }
   },
 };
 </script>
@@ -55,8 +79,6 @@ export default {
   justify-content: center; /* 水平居中 */
   height: 100%;
 }
-
-
 
 .title {
   font-size: 30px;
