+++ vue.config.js
@@ -1,6 +1,10 @@
 module.exports = {
   pluginOptions: {
     electronBuilder: {
+      // Use this to change the entrypoint of your app's main process
+      mainProcessFile: 'src/main/background.js',
+      // Use this to change the entry point of your app's render process. default src/[main|index].[js|ts]
+      rendererProcessFile: 'src/renderer/main.js',
       nodeIntegration: true,
       //  externals: ['!art-template'],
       // If you are using Yarn Workspaces, you may have multiple node_modules folders
